Added code to grab the data from the PatientInfo table for every appointment.

+++ Project-2-EMS/App_Code/Patient.cs
@@ -1,6 +1,6 @@
 ﻿namespace Project_2_EMS {
   internal class Patient {
-    public Patient(int patientId, string firstName, string lastName, string address, double balance) {
+    public Patient(int patientId, string firstName, string lastName, string address, decimal balance) {
       PatientId = patientId;
       FirstName = firstName;
       LastName = lastName;
@@ -12,6 +12,6 @@
     private string FirstName { get; }
     private string LastName { get; }
     private string Address { get; }
-    private double Balance { get; }
+    private decimal Balance { get; }
   }
 }
